% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aec_age_trends_plots.R
\name{aec_age_trend_pc}
\alias{aec_age_trend_pc}
\title{Graph trend in percentage of cases by age group}
\usage{
aec_age_trend_pc(collection, data, x, y, sex, group)
}
\arguments{
\item{collection}{One of CDI or Bacteraemia}

\item{data}{A data frame}

\item{x}{Variable giving financial year}

\item{y}{Variable giving percentage of cases the age group makes for all cases for year}

\item{sex}{Variable giving the sex group}

\item{group}{Variable giving age group (using mandatory age groupings)}
}
\value{
A ggplot2 object
}
\description{
Prepares a graph for the trend in the proportion of cases that are from a particular age group.
Intended to be used in conjunction with \code{\link{aec_age_trend_rate}} to plot a side-by-side figure.
See \code{\link{trends_data}} for example of format for data to be used with this function.
}
\examples{
data(age_trends_data)
p <- aec_age_trend_pc(collection = "CDI", data = age_trends_data,
    x = "fyear6", y = "age_sex_pc", sex = "sex", group = "age_group_new")
p
}
\seealso{
\code{\link{aec_age_trend_rate}}
#' @seealso \code{\link{trends_data}}
}
