% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aec_age_trends_plots.R
\name{aec_age_trend_rate}
\alias{aec_age_trend_rate}
\title{Graph trend in rate of cases by age and sex}
\usage{
aec_age_trend_rate(collection, data, x, y, sex, group, log_scale = FALSE)
}
\arguments{
\item{collection}{One of CDI or Bacteraemia}

\item{data}{A data frame}

\item{x}{Variable giving financial year}

\item{y}{Variable giving rate for age group by sex and year}

\item{sex}{Variable giving the sex group}

\item{group}{Variable giving age group (using mandatory age groupings)}

\item{log_scale}{Logical for whether a log y scale should be used. Defaults to FALSE}
}
\value{
A ggplot2 object
}
\description{
Prepares a graph for the trend in the rate of cases that are from a particular age group.
Intended to be used in conjunction with \code{\link{aec_age_trend_pc}} to plot a side-by-side figure.
See \code{\link{age_trends_data}} for example of format for data to be used with this function.
}
\examples{
data(age_trends_data)
q <- aec_age_trend_rate(collection = "CDI", data = age_trends_data,
    x = "fyear6", y = "rate", sex = "sex", group = "age_group_new",
    log_scale = TRUE)
q
}
\seealso{
\code{\link{aec_age_trend_pc}}

\code{\link{aec_age_trend_rate_pc_change}}
}
