% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aec_create_folder_structure.R
\name{aec_create_folder_structure}
\alias{aec_create_folder_structure}
\title{Set up empty folders for Annual publication}
\usage{
aec_create_folder_structure(fyear)
}
\arguments{
\item{fyear}{Financial year for report in YY_YY, e.g. 17_18 for 2017/2018 report}
}
\value{
New folder structure. Will print to console on success
}
\description{
Annual publication for mandatory surveillance should follow the same structure so that exports do not break.
This function creates the main folder on the K: drive and sets up empty subfolders.
}
\examples{
\dontrun{
aec_create_folder_structure("17_18")
}
}
