% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aec_ec_primary_focus_plot.R
\name{aec_ec_source_plot}
\alias{aec_ec_source_plot}
\title{Plot the most likely primary focus for E. coli bacteraemia by time to onset}
\usage{
aec_ec_source_plot(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A ggplot2 object
}
\description{
Although the function takes a single argument (a data frame),
it relies on the presence of several variables to that data frame:
* \code{tto_group} The grouped time to onset, generated by \code{\link{group_tto_bacteraemia}}
* \code{primary_focus} One of the primary foci in \code{\link{ecoli_primary_focus}}
* \code{pc} The percentage that the focus occurs within tto_group
}
\examples{
\dontrun{
dat <- structure(list(
tto_group = structure(c(1L, 1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L, 2L, 3L, 3L, 3L, 3L, 3L),
.Label = c("<2", "2-6", "ge7"), class = "factor"),
primary_focus = structure(c(1L, 2L, 3L, 4L, 5L, 1L, 2L, 3L, 4L, 5L, 1L, 2L, 3L, 4L, 5L),
.Label = c("gastro", "hepatobil", "other", "unknown", "uti"), class = "factor"),
pc = c(0.2, 0.2, 0.2, 0.2, 0.2, 0.5, 0.2, 0.2, 0.1, 0.3, 0.2, 0.2, 0.1, 0.1,
0.1)), .Names = c("tto_group", "primary_focus", "pc"),
class = "data.frame", row.names = c(NA, -15L))

p <- aec_ec_source_plot(dat)
p
}
}
\seealso{
\code{\link{group_tto_bacteraemia}} \code{\link{ecoli_primary_focus}}
}

