% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aec_inpatient.R
\name{aec_inpatient}
\alias{aec_inpatient}
\title{Determine whether record is from an acute trust inpatient for AEC}
\usage{
aec_inpatient(patient_location, patient_category, specimen_date,
  date_admitted)
}
\arguments{
\item{patient_location}{Location of patient at time of specimen.}

\item{patient_category}{Category of patient at time of specimen. One of:
"In-patient", "Day patient" or "Emergency Assessment"}

\item{specimen_date}{Date specimen taken}

\item{date_admitted}{Date of admission to NHS Acute Trust}
}
\value{
A numeric value
}
\description{
Inpatients have a patient location of 'NHS Acute Trust' and are one of:
inpatient, day patient or emergency assessment.
Assumes that where the date of admission is greater than the date of specimen, the patient was not actually an inpatient.
}
\examples{
patient_cat <- "In-patient"
patient_loc <- "NHS Acute Trust"
date_admitted <- lubridate::dmy("01/01/2016")
specimen_date <- lubridate::dmy("02/01/2016")
aec_inpatient(patient_loc, patient_cat, specimen_date, date_admitted)
\dontrun{
date_admitted <- NA
aec_inpatient(patient_loc, patient_cat, specimen_date, date_admitted)
date_admitted <- lubridate::dmy("03/01/2016")
aec_inpatient(patient_loc, patient_cat, specimen_date, date_admitted)
specimen_date <- NA
aec_inpatient(patient_loc, patient_cat, specimen_date, date_admitted)
}
}
