% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_trends_data}
\alias{age_trends_data}
\title{Trends in age of patients with *C. difficile* infection}
\format{
A data frame with 108 rows and 7 variables:
\describe{
  \item{fyear6}{Financial year}
  \item{age_group_new}{Mandatory surveillance age group}
  \item{n}{Count of cases for age group-year}
  \item{popn}{Count of population for England for age group-year}
  \item{rate}{Rate of CDI per 100,000 population}
  \item{age_sex_pc}{Per cent of cases for age group-year}
  \item{sex}{Sex}
}
}
\source{
\url{https://hcaidcs.phe.org.uk/WebPages/GeneralHomePage.aspx}
}
\usage{
age_trends_data
}
\description{
A data set containing data on trends in population rates of patients with
*C. difficile* infection in England, between 2007/08 and 2015/16
}
\keyword{datasets}
