% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_ccg_names.R
\name{ann_tab_ccg_names}
\alias{ann_tab_ccg_names}
\title{Format CCG names for printing in the annual tables.}
\usage{
ann_tab_ccg_names(x)
}
\arguments{
\item{A}{vector of CCG names}
}
\description{
This turns CCG names into title case format and removes strings 'NHS' and 'CCG'.
}
\examples{
x <- "NHS DARLINGTON CCG"
ann_tab_ccg_names(x)
}
