% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_long.R
\name{ann_tab_long}
\alias{ann_tab_long}
\title{Create long annual tables from line listing data}
\usage{
ann_tab_long(dat, collection_var, org_var, metric_var, period_var)
}
\arguments{
\item{dat}{A dataframe containing line-listing data}

\item{collection_var}{A string giving the data collection}

\item{org_var}{The variable giving the organisation code}

\item{metric_var}{Column in dat giving column for the metric}

\item{period_var}{Column giving the period by which counts should be aggregated}

\item{spec_date}{Column in dat giving specimen date}
}
\value{
A dataframe with counts of cases in long format, by organisation and time period
}
\description{
From FY 2018/19, annual tables will be moving to a long format rather than a
wide format. This function takes line-listing data and aggregates by
organisation and time period, producing counts of cases by a given metric
(apportioning or prior healthcare exposure).
}
\details{
The data are expanded to cover time periods in the data, but there may well
be organisations that are not in the line-listing data that will be merged in
later to make a complete table.

In addition, PHEC and organisation names will need to be merged in, as will
denominator data, and rates will need to be calculated.
}
\examples{
library(dplyr)
data(line_listing)
line_listing$fy_long <- fy_long(line_listing$specimen_date)

line_listing \%>\%
  filter(collection == "C. difficile") \%>\%
  ann_tab_long(dat = ., collection_var = collection,
    org_var = reporting_organisation_code,
    metric_var = prior_hc, period_var = fy_long)
}
