% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_onset.R
\name{ann_tab_onset_annual}
\alias{ann_tab_onset_annual}
\title{ann_tab_onset_annual}
\usage{
ann_tab_onset_annual(dat, timeperiod, org_code, denominator, total, ho, co,
  org_type)
}
\arguments{
\item{dat}{A dataframe in long format}

\item{timeperiod}{A financial year as YYYYYY or YYYY/YY. Can be either numeric or character}

\item{org_code}{Character vector giving organisation code}

\item{denominator}{The denominator value}

\item{total}{Total count of cases}

\item{ho}{Count of hospital-onset cases}

\item{co}{Count of community-onset cases}

\item{org_type}{The only parameter which should not be a column in the data frame a character giving either "Trust" or "CCG"}
}
\value{
A wide dataframe with total counts, total rates, onset count, onset rate, other onset count by financial year
}
\description{
Produce annual tables by onset status
}
\examples{
ann_onset_testdat <- data.frame(stringsAsFactors=FALSE,
              time_period = c(201415L, 201516L, 201617L, 201718L, 201415L, 201516L,
                              201617L, 201718L),
                 org_code = c("W1A", "W1A", "W1A", "W1A", "E17", "E17", "E17",
                              "E17"),
              denominator = c(100L, 100L, 100L, 100L, 100L, 100L, 100L, 100L),
                    this_is_the_total = c(10L, 20L, 30L, 40L, 50L, 0L, 40L, 30L),
                       ho = c(8L, 16L, 22L, 31L, 40L, 0L, 35L, 15L),
                       co = c(2L, 4L, 8L, 9L, 10L, 0L, 5L, 15L),
       pir_trust_assigned = c(4L, 10L, 5L, 30L, 10L, 0L, 5L, 5L),
         pir_ccg_assigned = c(3L, 5L, 20L, 5L, 40L, 0L, 30L, 20L),
              third_party = c(3L, 5L, 5L, 5L, 0L, 0L, 5L, 5L)
    )
 ann_tab_onset_annual(ann_onset_testdat, timeperiod = time_period,
    org_code = org_code, denominator = denominator, total= this_is_the_total, ho = ho,
    co = co, org_type = "trust")
}
