% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_pir_org_type.R
\name{ann_tab_pir_org_type}
\alias{ann_tab_pir_org_type}
\title{Format pir org type for annual tables}
\usage{
ann_tab_pir_org_type(x)
}
\arguments{
\item{x}{A string vector giving the final pir assignment type}
}
\value{
A factor variable with four levels: Total Reported Cases,
Trust Assigned Cases, CCG Assigned Cases and Third Party Cases
}
\description{
Format pir org type for annual tables
}
\examples{
x <- "NHS Trust"
ann_tab_pir_org_type(x)
}
