% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportion.R
\name{apportion}
\alias{apportion}
\title{Apportion CDI and MSSA records}
\usage{
apportion(collection, patient_location, patient_category, date_admitted,
  specimen_date, date_entered)
}
\arguments{
\item{collection}{The collection (MRSA, MSSA, CDI, E. coli) to which the record belongs.}

\item{patient_location}{The patient's location at time of sample}

\item{patient_category}{The patient's category at time of sample}

\item{date_admitted}{The date the patient was admitted for records from inpatients}

\item{specimen_date}{The date of the specimen (POSIXct, POSIXt or Date format)}

\item{date_entered}{Date record created normally 'data collection date' (POSIXct, POSIXt or Date format)}
}
\value{
A binary vector giving 1 for Trust-apportioned records and 0 for non-Trust apportioned records
}
\description{
Apportions records to apportioning algorithm.
Includes a clause whereby records created prior to 26th October 2015 are not trust apportioned if patient location is Unknown.
This is to ensure consistency with previous reports. See p 47 of Mandatory Surveillance protocol.
Prior to the launch of the new DCS, users could create records where the patient location was Null and these cases would not have been apportioned automatically.
}
\examples{
apportion("mssa", "NHS Acute Trust", "In-patient",
date_admitted = lubridate::dmy("01-01-2015"), specimen_date =
lubridate::dmy("05-01-2015"), date_entered = lubridate::dmy("26-10-2015"))
\dontrun{
apportion("mssa", "NHS Acute Trust", "In-patient",
date_admitted = as.Date(NA_real_, origin = "01-01-1970"),
specimen_date = lubridate::dmy("01-01-2015"), date_entered =
lubridate::dmy("26-10-2015"))
apportion("mssa", "NHS Acute Trust", "In-patient", date_admitted =
lubridate::dmy(""), specimen_date = lubridate::dmy("01-01-2015"),
date_entered = lubridate::dmy("26-10-2015"))
}
apportion("mssa", "NHS Acute Trust", "In-patient", date_admitted =
lubridate::dmy("01-01-2015"),  specimen_date = lubridate::dmy("01-01-2015"),
date_entered = lubridate::dmy("26-10-2015"))
apportion("mssa", "NHS Acute Trust", "Outpatient", date_admitted =
lubridate::dmy("01-01-2015"),  specimen_date = lubridate::dmy("05-01-2015"),
date_entered = lubridate::dmy("26-10-2015"))
apportion("cdi", "NHS Acute Trust", "In-patient", date_admitted =
lubridate::dmy("01-01-2015"),  specimen_date = lubridate::dmy("05-01-2015"),
date_entered = lubridate::dmy("26-10-2015"))
}
