% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportion_ecoli.R
\name{apportion_ecoli}
\alias{apportion_ecoli}
\title{Provisional algorithm for apportioning *E. coli* BSI to NHS Trusts or not}
\usage{
apportion_ecoli(patient_location, patient_category, date_admitted,
  specimen_date, date_entered)
}
\arguments{
\item{patient_location}{The patient's location at time of sample}

\item{patient_category}{The patient's category at time of sample}

\item{date_admitted}{The date the patient was admitted for records from inpatients}

\item{specimen_date}{The date of the specimen (POSIXct, POSIXt or Date format)}

\item{date_entered}{Date of admission (POSIXct, POSIXt or Date format)}

\item{collection}{The collection (MRSA, MSSA, CDI, E. coli) to which the record belongs.}
}
\value{
An ordered factor giving non-Trust; Trust, 48 hours to 7 days; Trust, > 7 days or Trust, no date of admission.
}
\description{
Based on apportionment algorithm for MSSA and CDI, but with two options for Trust-apportioned cases.
p48 of \href{https://hcaidcs.phe.org.uk/ContentManagement/LinksAndAnnouncements/HCAIDCS_Mandatory_Surveillance_Protocol_v4.0.pdf}{the mandatory surveillance protocol}.
}
\examples{
dat <- structure(list(
  patient_location = c("NHS Acute Trust", "NHS Acute Trust", "NHS Acute Trust", 
                       "NHS Acute Trust", "NHS Acute Trust", "", "", "", "", "",
                       "NHS Acute Trust", "NHS Acute Trust", "NHS Acute Trust", 
                       "NHS Acute Trust", "NHS Acute Trust", "", "", "", "", "", 
                       "NHS Acute Trust", "", "Nursing Home", "Nursing Home", 
                       "NHS Acute Trust", "NHS Acute Trust", "NHS Acute Trust", 
                       "Nursing Home"), 
  patient_category = c("In-patient", "Day patient", "Emergency Assessment", 
                       "Unknown", "", "In-patient", "Day patient", 
                       "Emergency Assessment", "Unknown", "", "In-patient", 
                       "Day patient", "Emergency Assessment", "Unknown", "", 
                       "In-patient", "Day patient", "Emergency Assessment", 
                       "Unknown", "", "A&E only", "A&E only", "In-patient", 
                       "In-patient", "In-patient", "Day patient", 
                       "Emergency Assessment", "In-patient"), 
  date_admitted = structure(c(16749, 16749, 16749, 16749, 16749, 16749, 16749, 
                              16749, 16749, 16749, 16749, 16749, 16749, 16749, 
                              16749, 16749, 16749, 16749, 16749, 16749, 16749, 
                              16749, 16749, 16749, 16749, NA, NA, NA), 
                            class = "Date"), 
  specimen_date = structure(c(16749, 16751, 16760, 16751, 16751, 16749, 16751, 
                              16760, 16751, 16751, 16749, 16751, 16760, 16751, 
                              16751, 16749, 16751, 16760, 16751, 16751, 16749, 
                              16749, 16749, 16749, 16749, 16751, 16760, 16749), 
                            class = "Date"), 
  date_entered = structure(c(16765, 16765, 16765, 16765, 16765, 16765, 16765, 
                             16765, 16765, 16765, 16728, 16728, 16728, 16728, 
                             16728, 16728, 16728, 16728, 16728, 16728, 16765, 
                             16728, 16765, 16765, 16765, 16765, 16765, 16765), 
                           class = "Date")), class = "data.frame", 
  .Names = c("patient_location", "patient_category", "date_admitted
             ", "specimen_date", "date_entered"), row.names = c(NA, -28L))
 dat \%>\% mutate(apportioned = apportion_ecoli(patient_location, 
 patient_category, date_admitted, specimen_date, date_entered))
 @export
}

