% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_team_names.R
\name{area_team_names}
\alias{area_team_names}
\title{Format CCG and NHS area team names into print format.}
\usage{
area_team_names(area_team_name)
}
\arguments{
\item{area_team_name}{A vector of area team names}
}
\description{
Removes brackets and sorts capitalisation.
}
\examples{
x <- "NHS ENGLAND NORTH (YORKSHIRE AND HUMBER)"
area_team_names(x)
}
\seealso{
\code{\link{ann_tab_ccg_names}}
}

