% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assignment_data2}
\alias{assignment_data2}
\title{Data for testing assignement algorithm}
\format{A data frame with 15 rows of 7 variables:
\describe{
 \item{pircasestatus}{string}
 \item{assignmentmethodcode}{Numeric giving assignment method}
 \item{patientlocation}{string}
 \item{provisionalorganisationname}{string}
 \item{finalpirassignedorganisation}{string}
 \item{patient_category}{string}
 \item{id}{numeric}
}}
\usage{
assignment_data2
}
\description{
A sample data set for testing the assignment_algorithm function.
}
\keyword{datasets}
