% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{at_sp_df}
\alias{at_sp_df}
\title{NHS Area Team geographies for England}
\format{A data frame with ~87.5k observations and 12 columns:
\describe{
 \item{long}{Longitude}
 \item{lat}{Lattitude}
 \item{group}{Group id for polygons}
 \item{id}{Numeric id for Area Team, 0-9}
 \item{NHSATCD}{ONS code for Area Team}
 \item{NHSATNM}{Area Team name}
}}
\source{
"Office for National Statistics"
}
\usage{
at_sp_df
}
\description{
A dataframe containing the geographies for NHS Area Teams.
}
\keyword{datasets}

