% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_days_fq.R
\name{bed_days_fq}
\alias{bed_days_fq}
\title{Calculates inpatient bed-days from Actue Trust KH03 returns}
\usage{
bed_days_fq(data, kh03, qtr)
}
\arguments{
\item{data}{A dataframe}

\item{kh03}{Trust specific KH03 mandatory return}

\item{qtr}{Financial quarter. Example format; 20171 (Apr-Jun 2017), 20172 (Jul-Sep 2017), 20173 (Oct-Dec 2017) and 20172 (Jan-Mar 2018)}
}
\value{
Dataframe with the number of days in each quarter with corresponding inpatient bed-days for the same period
}
\description{
This package calculates NHS Acute Trusts's inpatient bed-days using NHSI's published mandatory KH03 return
}
\examples{
average_kh03 <- data.frame(
fq = c(20171,20172,20173,20174,20151,20152,20153,20154),
trust_code = c(rep("RX1",4), rep("RYJ",4)),
trust_name = c(rep("Nottingham University Hospitals",4), rep("Imperial College Healthcare",4)),
kh03_avg = c(1425,1357,1392,1481,965,940,933,970)
)

bed_days_fq(average_kh03,"kh03_avg","fq")

}
