% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_days_fy.R
\name{bed_days_fy}
\alias{bed_days_fy}
\title{Calculates inpatient bed-days from Actue Trust KH03 returns}
\usage{
bed_days_fy(data, kh03, fyear)
}
\arguments{
\item{data}{A dataframe}

\item{kh03}{Trust specific KH03 mandatory return}

\item{fy}{Financial year. Example format; 201718 (Apr 2017 to March 2018) and 201819 (Apr 2018 to March 2019)}
}
\value{
Dataframe with the number of days in each financial year with corresponding inpatient bed-days for the same period
}
\description{
This package calculates NHS Acute Trusts's inpatient bed-days using NHSI's published mandatory KH03 return
}
\examples{
average_kh03 <- data.frame(
fy = c(200708,200809,200708,200809),
trust_code = c(rep("RX1",2), rep("RYJ",2)),
trust_name = c(rep("Nottingham University Hospitals",2), rep("Imperial College Healthcare",2)),
kh03_avg = c(1508,1537,1211,1250)
)

bed_days_fy(average_kh03,"kh03_avg","fy")

}
