% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_q_long.R
\name{cal_q_long}
\alias{cal_q_long}
\title{Format date into calendar year quarter.}
\usage{
cal_q_long(date_var)
}
\arguments{
\item{date_var}{A date variable in class Date or POSIX}
}
\value{
A text string giving the quarter and calendar year.
}
\description{
Formats date from date or POSIX class to a text string giving the calendar
quarter
}
\examples{
x <- lubridate::dmy("01/01/2001")
fq_long(x)
}
\seealso{
\code{\link{fy_long}}
}

