% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_class.R
\name{check_date_class}
\alias{check_date_class}
\title{Check that date variables are in required format}
\usage{
check_date_class(date_var)
}
\arguments{
\item{date_var}{A date variable to be checked.}
}
\description{
Checks that dates are in class date or POSIX. Called by other functions.
}
