% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_console.R
\name{clear_console}
\alias{clear_console}
\title{Clear the console}
\usage{
clear_console()
}
\description{
Clears the console so that you can start the next step of coding with a fresh screen.
Has no arguments.
}
\examples{
cat(rep("This is a line. \\n", 50)) # print some text to clear
clear_console()
}

