% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmonth_to_fmonth.R
\name{cmonth_to_fmonth}
\alias{cmonth_to_fmonth}
\title{cmonth_to_fmonth}
\usage{
cmonth_to_fmonth(numeric_month)
}
\arguments{
\item{numeric_month}{A numeric value for a calendar month where January = 1}
}
\value{
A numeric value where April = 1 and January = 10
}
\description{
Converts calendar months to financial year months
}
\examples{
cmonth_to_fmonth(1) # January
cmonth_to_fmonth(2) # February
cmonth_to_fmonth(4) # April
cmonth_to_fmonth(12) # December
}
