% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_to_string.R
\name{cmonth_to_string}
\alias{cmonth_to_string}
\title{cmonth_to_string}
\usage{
cmonth_to_string(x)
}
\arguments{
\item{x}{A numeric monthly value where 1 = January}
}
\value{
A string giving month name
}
\description{
Convert numeric calendar month to string.
Mostly important for monthly reports
}
\examples{
cmonth_to_string(1)
cmonth_to_string(12)
cmonth_to_string(13)
}
