% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dcs.R
\name{collapse_select}
\alias{collapse_select}
\title{collapse_select}
\usage{
collapse_select(cols)
}
\arguments{
\item{cols}{vector output of \code{dcs_cols_lookup}}
}
\description{
collapse output from \code{dcs_cols_lookup} into the start of a select query
}
\examples{
cols <- dcs_cols_lookup(cols = c("reporting_organisation_code", "ccg"))
collapse_select(cols)
}
