% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_csv2r.R
\name{convert_csv2R}
\alias{convert_csv2R}
\title{Convert .csv files to global variables that can be used in merge_everything function.}
\usage{
convert_csv2R(trust_file, ccg_file)
}
\arguments{
\item{trust_file}{A file with CCG codes.}

\item{ccg_file}{A file with CCG codes.}
}
\value{
R objects (lists of list) to be used in merge_everything function.
}
\description{
Create trust_list and ccg_list objects.
}
