% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_csv2r.R
\name{convert_csv2RCCG}
\alias{convert_csv2RCCG}
\title{Convert .csv file to global variable that can be used in merge_everything function.}
\usage{
convert_csv2RCCG(ccg_file)
}
\arguments{
\item{ccg_file}{A file with CCG codes.}
}
\value{
R object(list).
}
\description{
Creates ccg_list objects.
}
