% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_csv2r.R
\name{convert_csv2RTrust}
\alias{convert_csv2RTrust}
\title{Convert .csv file to global variables that can be used in merge_everything function.}
\usage{
convert_csv2RTrust(trust_file)
}
\arguments{
\item{trust_file}{A file with CCG codes.}
}
\value{
R object(list).
}
\description{
Creates trust_list objects.
}
