% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyear_to_cq.R
\name{cyear_to_cq}
\alias{cyear_to_cq}
\title{Cyear to long calendar quarter}
\usage{
cyear_to_cq(cyear)
}
\arguments{
\item{cyear}{A numeric or string variable giving the calendar year and quarter, e.g. 20112 - July-September 2011}
}
\value{
A text string giving the quarter and then the year.
}
\description{
Converts a five digit financial year into a long text string
}
\examples{
cyear_to_cq(20112)
}
\seealso{
\code{\link{kh03_year}}, \code{\link{fyear_to_fq}}
}
