% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyear_to_fyear.R
\name{cyear_to_fyear}
\alias{cyear_to_fyear}
\title{Convert calendar year to financial year}
\usage{
cyear_to_fyear(x)
}
\arguments{
\item{x}{Numeric year}
}
\value{
Character string giving the six character financial year
}
\description{
Convert calendar year to financial year
}
\examples{
cyear_to_fyear(2012)
}

