% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_compare.R
\name{data_compare}
\alias{data_compare}
\title{Compare two data frames to identify cells which are different}
\usage{
data_compare(currentdf, lastdf)
}
\arguments{
\item{currentdf}{A data frame in wide format giving the values for the current publication.}

\item{lastdf}{A data frame in wide format giving the values for the most recent publication prior to the one being prepared.}
}
\value{
A matrix specifying cells that have changed.
}
\description{
Used to identify cells which have changed between publications for later highlighting in Excel export.
With thanks to Amy Mikhail for writing this function for me.
Not currently production.
}
