% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_end_period.R
\name{date_end_period}
\alias{date_end_period}
\title{Get the string of the first date in a period}
\usage{
date_end_period(x, period_type)
}
\arguments{
\item{x}{A period, either string or numeric}

\item{period_type}{A string, one of "fq", "cq" or "fyear6"}
}
\value{
A \strong{string} in format dd/mm/yyyy
}
\description{
Takes a period used in mandatory surveillance data and returns first date in the period.
}
\examples{
date_end_period(20154, "fq")
date_end_period(20154, "cq")
date_end_period(201516, "fyear6")
dat <- data.frame(x = c(20151, 20152, 20153, 20154))
date_end_period(dat$x, "fq")
date_end_period(dat$x, "cq")
}
\seealso{
\code{\link{date_start_period}}
}
