% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_cmonth.R
\name{date_to_cmonth}
\alias{date_to_cmonth}
\title{date_to_cmonth}
\usage{
date_to_cmonth(x, as_factor = FALSE)
}
\arguments{
\item{x}{A date}

\item{as_factor}{TRUE or FALSE, defaults to FALSE}
}
\value{
Either a numeric value giving the month of the year (1 = January) or a factor with the name of the month of the year.
}
\description{
Convert a date to a calendar month, either as a numeric value, or as an ordered
factor.
}
\examples{

x <- as.Date("01/01/2019", format = "\%d/\%m/\%Y")
date_to_cmonth(x)
date_to_cmonth(x, as_factor = TRUE)
}
