% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_fmonth.R
\name{date_to_fmonth}
\alias{date_to_fmonth}
\title{date_to_fmonth}
\usage{
date_to_fmonth(x)
}
\arguments{
\item{x}{a date value}
}
\value{
a numeric value for the month of financial year, e.g. 1 for April, and 12 for March
}
\description{
Convert dates to months of financial year
}
\examples{
date_to_fmonth(lubridate::dmy("01-01-2018"))
date_to_fmonth(lubridate::dmy("01-02-2018"))
date_to_fmonth(lubridate::dmy("01-03-2018"))
date_to_fmonth(lubridate::dmy("01-04-2018"))
date_to_fmonth(lubridate::dmy("01-12-2018"))
}
