% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_fq.R
\name{date_to_fq}
\alias{date_to_fq}
\title{date_to_fq}
\usage{
date_to_fq(x)
}
\arguments{
\item{x}{A date}
}
\value{
A numeric value giving YYYYQ
}
\description{
Convert a date to a five-character financial quarter
}
\examples{
x <- as.Date("01/01/2017", format = "\%d/\%m/\%Y")
date_to_fq(x)
}
