% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_fyear.R
\name{date_to_fyear}
\alias{date_to_fyear}
\title{date_to_fyear}
\usage{
date_to_fyear(x)
}
\arguments{
\item{x}{A date object}
}
\value{
A numeric value giving the four-digit financial year
}
\description{
Convert a date object to a four-digit financial year
}
\examples{
x <- as.Date("01/01/2017", format = "\%d/\%m/\%Y")
date_to_fyear(x)
}
