% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_cyear.R
\name{days_in_cy}
\alias{days_in_cy}
\title{Get the number of days in a calendar year}
\usage{
days_in_cy(x)
}
\arguments{
\item{x}{A calendar year in format yyyy}
}
\value{
The number of days in the calendar year
}
\description{
Need to obtain the number of days in a year
}
\examples{
days_in_cy(2012)
days_in_cy(2011)
}
