% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_fmonth.R
\name{days_in_fmonth}
\alias{days_in_fmonth}
\title{Calculates number of days in a financial month.}
\usage{
days_in_fmonth(month_no, fyear)
}
\arguments{
\item{month_no}{numeric month of financial year, where 1 = April}

\item{fyear}{4 character financial year where 2012 is year 2012/13}
}
\value{
numeric value for number of days in month
}
\description{
Used for calculating denominators in monthly scripts.
}
\examples{
days_in_fmonth(1, 2017)
x <- days_in_fmonth(1, 2017)
class(x)
days_in_fmonth(11, 2012) # February on a leap year
days_in_fmonth(11, 2013) # February on an ordinary year
}
