% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_fq.R
\name{days_in_fq}
\alias{days_in_fq}
\title{Get the number of days in the quarter of a financial year}
\usage{
days_in_fq(x)
}
\arguments{
\item{x}{A financial year quarter in format "yyyyq"}
}
\value{
The number of days in the quarter, accounting for leap years.
}
\description{
For working with KH03, quarters are in format "yyyyq".
Want number of days per quarter and this varies by leap year.
This method takes leap year into account.
}
\details{
\strong{Note:} This function must be used in conjunction with \code{\link[dplyr]{rowwise}} from dplyr or \code{\link[base]{sapply}}.
}
\examples{
days_in_fq("20131")
days_in_fq("20134")
# 2016 is a leap year, should have 91 days in quarter 1 (i.e. 20154)
days_in_fq("20154")
x <- data.frame(fyear = c(20151, 20152, 20153, 20154))
x$days <- days_in_fq(x$fyear)
}

