% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_in_fy.R
\name{days_in_fy}
\alias{days_in_fy}
\title{Get the number of days in a financial year}
\usage{
days_in_fy(x)
}
\arguments{
\item{x}{A financial year in format "y1y1y1y1y2y2"}
}
\value{
The number of days in the financial year, from 1st April to 31st March, as a numeric value.
}
\description{
For working with KH03, years between 2007/08 and 2009/10 are in format "y1y1y1y1y2y2".
Need to obtain the number of days in a year
}
\examples{
days_in_fy("200708")
}

