% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dcs.R
\name{dcs_cols_lookup}
\alias{dcs_cols_lookup}
\title{dcs_cols_lookup}
\usage{
dcs_cols_lookup(cols)
}
\arguments{
\item{cols}{A vector of desired column names}
}
\description{
convert desired column names into DCS database column names, with table prefix
}
\details{
An internal function that provides a vectorised lookup of column names.
Non-matching column names are converted to NA and filtered out
}
\examples{
dcs_cols_lookup(cols = c("reporting_organisation_code", "ccg"))
}
