% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcs_denom_boilerplate.R
\name{dcs_popn_boilerplate}
\alias{dcs_popn_boilerplate}
\title{Create population boilerplate columns for DCS denominator upload}
\usage{
dcs_popn_boilerplate(x)
}
\arguments{
\item{x}{a data frame to which columns will be added}
}
\value{
A data frame with denominator_code, denominator_descriptionnotrequd,
denominator_type_code, range_from, range_to and per_factor
}
\description{
Adds columns that are common between all organisation types when preparing denominators for upload to DCS
}
