% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcs_specialty_grouping.R
\name{dcs_speciality_grouping}
\alias{dcs_speciality_grouping}
\title{Group specialties from the HCAI DCS}
\usage{
dcs_speciality_grouping(spec_nm)
}
\arguments{
\item{df}{Character. Specialty name}
}
\description{
Group specialties from the HCAI DCS
}
\examples{

dcs_speciality_grouping("Clinical oncology (radiotherapy)")

dcs_speciality_grouping("PIC Gastroenterology")

dcs_speciality_grouping("PIC Nephrology")

dcs_speciality_grouping("")

dcs_speciality_grouping("  ")

dcs_speciality_grouping(NA)

dcs_speciality_grouping("xds")


}
