% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_hcaidcs.R
\name{detach_hcaidcs}
\alias{detach_hcaidcs}
\title{Convenience function to detach hcaidcs package}
\usage{
detach_hcaidcs()
}
\description{
I often need to detach this package to reload an updated version, this should do so nice and quickly.
I don't see other uses for it.
Inspiration for this function: \href{http://stackoverflow.com/questions/6979917/how-to-unload-a-package-without-restarting-r}{kohske on StackOverflow.com}
}
