% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_apportion_step1.R
\name{ec_apportion_step1}
\alias{ec_apportion_step1}
\title{for ecoli apportioning}
\usage{
ec_apportion_step1(patient_location, date_entered)
}
\arguments{
\item{patient_location}{The patients location at time of sample}

\item{date_entered}{Date format the record was created on the system}
}
\description{
for ecoli apportioning
}

