% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoli_primary_focus.R
\name{ecoli_primary_focus}
\alias{ecoli_primary_focus}
\title{Classify the source of bacteramia for E. coli}
\usage{
ecoli_primary_focus(x)
}
\arguments{
\item{x}{Character variable giving the most likely primary focus of bacteraemia}
}
\value{
An ordered factor giving the grouped primary focus
}
\description{
Groups the most likely primary focus of E. coli bacteraemia according to AEC table.
}
\examples{
ecoli_primary_focus("Urinary tract")
ecoli_primary_focus("Central nervous system")
ecoli_primary_focus("")
}
