% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_apportioned.R
\name{factor_apportion_both}
\alias{factor_apportion_both}
\title{factor_apportion_both}
\usage{
factor_apportion_both(x)
}
\arguments{
\item{x}{A character variable from an apportioning algorithm}
}
\value{
An ordered factor variable
}
\description{
Converts apportioning data into combined factor for use in tables containing both onset status and prior healthcare exposure
}
\examples{
factor_apportion_both(1)
factor_apportion_both("HO")
factor_apportion_both("hoha")
}
\seealso{
\code{\link{mo_tab_long}}, \code{\link{ann_tab_long}}, \code{\link{factor_prior_hc}} and \code{\link{factor_apportioned}}
}
