% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_apportioned.R
\name{factor_prior_hc}
\alias{factor_prior_hc}
\title{factor_prior_hc}
\usage{
factor_prior_hc(x)
}
\arguments{
\item{x}{A character vector limited to hoha, coha, coia, coca, unknown_3_mo and all_blank}
}
\value{
An ordered factor variable
}
\description{
Converts data with prior healthcare exposure information into factor for
production of long format tables.
}
\examples{
factor_prior_hc("hoha")
}
\seealso{
\code{\link{mo_tab_long}}, \code{\link{ann_tab_long}}
}
