% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_next_pub_date.R
\name{first_wednesday}
\alias{first_wednesday}
\title{Returns the first Wednesday of a month
Used to calculated the next publication date of monthly outputs.}
\usage{
first_wednesday(x)
}
\arguments{
\item{x}{A date in Year, Month, Day format}
}
\value{
A date object
}
\description{
Returns the first Wednesday of a month
Used to calculated the next publication date of monthly outputs.
}
\examples{
first_wednesday(lubridate::floor_date(lubridate::today(), "month"))
\dontrun{
first_wednesday(lubridate::floor_date(today(), "month") \%m+\%
    lubridate::months(1))
    }
}
