% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_to_string.R
\name{fmonth_to_string}
\alias{fmonth_to_string}
\title{fmonth_to_string}
\usage{
fmonth_to_string(x)
}
\arguments{
\item{x}{A numeric monthly value where 1 = April}
}
\value{
A string giving month name
}
\description{
Convert numeric FY month to string.
Mostly important for monthly reports
}
\examples{
fmonth_to_string(1)
fmonth_to_string(12)
}
