% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq5.R
\name{fq5}
\alias{fq5}
\title{Convert a date to five-character financial quarter}
\usage{
fq5(x)
}
\arguments{
\item{x}{A date variable}
}
\value{
A numeric value with five digits giving the financial year in the first four digits and the quarter in the final digit.
}
\description{
Takes a date and returns a five character financial year and quarter (i.e. yyyyq).
}
\examples{
x <- lubridate::dmy("01/01/2012")
fq5(x)
x <- lubridate::dmy("01/04/2012")
fq5(x)
}

