% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_long.R
\name{fq_long}
\alias{fq_long}
\title{Format date into long quarter.}
\usage{
fq_long(date_var)
}
\arguments{
\item{date_var}{A date variable in class Date or POSIX}
}
\value{
A text string giving the quarter and year.
}
\description{
Formats date from date or POSIX class to a text string giving the the quarter
as 'January to March yyyy'
}
\examples{
x <- lubridate::dmy("01/01/2001")
fq_long(x)
}
\seealso{
\code{\link{fy_short}}

\code{\link{fy_long_short}}

\code{\link{ordered_fin_qtr}}
}
