% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_short.R
\name{fq_short}
\alias{fq_short}
\title{Format date into short text financial quarter.}
\usage{
fq_short(date_var)
}
\arguments{
\item{date_var}{A date variable in class Date or POSIX}
}
\value{
A text string giving the financial quarter, where April to June is
quarter 1 and January to March is quarter 4.
}
\description{
Formats date from date or POSIX class to a text string giving the financial
quarter as integer from 1 to 4.
}
\examples{
x <- lubridate::dmy("01/01/2001")
fq_short(x)
}
