% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_short.R
\name{fq_short_to_date}
\alias{fq_short_to_date}
\title{fq_short_to_date}
\usage{
fq_short_to_date(x)
}
\arguments{
\item{x}{A short financial year quarter e.g. 20171}
}
\value{
The first day of the first month of the financial year quarter
}
\description{
Convert short financial quarter to date
}
\examples{
fq_short_to_date(20124)
}
\seealso{
\code{\link{fq_short}}
}
