% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_to_cq.R
\name{fq_to_cq}
\alias{fq_to_cq}
\title{Convert financial year quarters to calendar year quarters}
\usage{
fq_to_cq(x)
}
\arguments{
\item{x}{A financial quarter, e.g. 20154 is January-March 2016 (cyear 20161)}
}
\value{
A numeric value giving calendar quarter in format `yyyyq`
}
\description{
Converts from five-character financial year quarter to five character calendar year quarter
}
\examples{
fq_to_cq(20154)
fq_to_cq(20151)
}

