% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_to_fyear.R
\name{fq_to_fyear}
\alias{fq_to_fyear}
\title{Convert financial quarter to six-character financial year}
\usage{
fq_to_fyear(x)
}
\arguments{
\item{x}{Either string or numeric value financial year/quarter, e.g. 20151 = April-June 2015}
}
\value{
Numeric value for financial year, e.g. 201516
}
\description{
Takes financial quarter in format yyyyq and returns yyyy/yy format financial year.
}
\examples{
fq_to_fyear(20151)
}

