% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy_long.R
\name{fy_long}
\alias{fy_long}
\title{Format date into long text financial year.}
\usage{
fy_long(date_var)
}
\arguments{
\item{date_var}{A date variable in class Date or POSIX}
}
\value{
A text string giving the financial year.
}
\description{
Formats date from date or POSIX class to a text string giving the financial
year as 'April yyyy to March yyyy'
}
\examples{
x <- lubridate::dmy("01/01/2001")
fy_long(x)
}
\seealso{
\code{\link{fy_short}}

\code{\link{fy_long_short}}

\code{\link{ordered_fin_qtr}}
}

