% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy_long_short.R
\name{fy_long_short}
\alias{fy_long_short}
\title{Format fy_long into short financial year.}
\usage{
fy_long_short(x)
}
\arguments{
\item{x}{A string variable giving long financial year created using
\code{\link{fy_long}}}
}
\value{
A text string giving the financial year.
}
\description{
Formats date from long financial as 'yyyyyy+1'.
Used for linking HCAI DCS data to formatted kh03 data.
}
\examples{
x <- lubridate::dmy("01/01/2001")
x <- fy_long(x)
fy_long_short(x)
}
\seealso{
\code{\link{fy_long}}
}
