% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy_six.R
\name{fy_six}
\alias{fy_six}
\title{Create six character financial year from date}
\usage{
fy_six(x)
}
\arguments{
\item{x}{A date character}
}
\value{
Six character string giving the financial year, e.g. 201213
}
\description{
Takes date variables and returns six character financial years (e.g. 201213) for merging with kh03 and other data.
Needs further work to return ordered factor.
}
\examples{
x <- lubridate::dmy("01-01-2012")
fy_six(x)
}
