% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy_long.R
\name{fy_six_to_long}
\alias{fy_six_to_long}
\title{Format date into long text financial year.}
\usage{
fy_six_to_long(x)
}
\arguments{
\item{x}{Either a numeric or character variable giving the financial year as
YYYYYY or YYYY/YY}
}
\value{
A text string giving the financial year.
}
\description{
Converts six-character financial year to long financial
year as 'April yyyy to March yyyy'.
Assumes that all dates will be >= 2000
}
\examples{
x <- 201718
fy_six_to_long(x)
x <- "2017/18"
fy_six_to_long(x)
}
\seealso{
\code{\link{fy_short}}

\code{\link{fy_long_short}}

\code{\link{ordered_fin_qtr}}

\code{\link{fy_long}}
}
