% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyear_to_cyear.R
\name{fyear_to_cyear}
\alias{fyear_to_cyear}
\title{Convert financial year to calendar year}
\usage{
fyear_to_cyear(x)
}
\arguments{
\item{x}{Numeric financial year}
}
\value{
Character string giving the four digit calendar year
}
\description{
Intended for the production of the fingertips annual data.
In fingertips the current year is the first year of the financial year.
i.e. for data from financial year 2015/16, the year is 2015.
It's worth noting that R will sort numbers stored as characters well.
}
\examples{
fyear_to_cyear(201213)
}
\seealso{
\code{\link{cyear_to_fyear}}
}
