% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyear_to_fq.R
\name{fyq_to_fq}
\alias{fyq_to_fq}
\title{Fyear-quarter to long financial quarter}
\usage{
fyq_to_fq(fyear)
}
\arguments{
\item{fyear}{A numeric or string variable giving the finacial year and quarter, e.g. 20112 - July-September 2011}
}
\value{
A text string giving the quarter and then the year.
}
\description{
Converts a five digit financial year and quarter into a long text string
}
\examples{
fyear_to_fq(20112)
}
\seealso{
\code{\link{kh03_year}}, \code{\link{cyear_to_cq}}
}
