% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_source_bacteraemia.R
\name{group_source_bacteraemia}
\alias{group_source_bacteraemia}
\title{group_source_bacteraemia}
\usage{
group_source_bacteraemia(collection, source_col)
}
\arguments{
\item{collection}{A string giving the data collection, one of
"MRSA", "MSSA", "E. coli", "Klebsiella spp", "Pseudomonas aeruginosa"}

\item{source_col}{The string giving the source of bacteraemia to be grouped}
}
\value{
A vector of grouped sources
}
\description{
Group the sources of bacteraemia for reporting in QEC or elsewhere.
In 2017 the column which gave the primary focus for Gram-negative collections
changed with the introducion  new risk factor collection on the DCS.
}
\details{
Previously, the field name was risk_factors_most_likely_primary_focus
the new field name is source_primary_focus_of_bacteraemia

 The general approach when analysing the data is to combine the information
from both fields into one.
This may cause problems where there is contradictory information.
New information is copied into old, but only when new information is not
unknown or otherwise missing.
}
\examples{
group_source_bacteraemia(collection = "E. coli", source_col = "Urinary tract")
\dontrun{
dat <- data.frame(stringsAsFactors=FALSE,
  org = c("E. coli", "MRSA"),
  source = c("Unassessed", "Skin/Soft tissue infection")
)
dat$grouped_source <- group_source_bacteraemia(collection = dat$org, source_col = dat$source)
}
}
