% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tto_bacteraemia.R
\name{group_tto_bacteraemia}
\alias{group_tto_bacteraemia}
\title{Provides grouped time to onset for bacteraemia cases}
\usage{
group_tto_bacteraemia(time_diff)
}
\arguments{
\item{time_diff}{The time to onset as calculated by \code{\link{time_to_onset}}}
}
\value{
An ordered factor
}
\description{
Provides grouped time to onset for bacteraemia cases
}
\examples{
onset <- dmy("05-01-2011")
admitted <- dmy("01-01-2011")
patient_loc <- "In-patient"
tto <- time_to_onset(admitted, onset, patient_loc)
group_tto_bacteraemia(tto)
}
\seealso{
\code{\link{time_to_onset}}

\code{\link{group_tto_cdi}}
}

