% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_compare.R
\name{highlight}
\alias{highlight}
\title{Compare workbooks for differences and highlight in red cells that are different.}
\usage{
highlight(old_file, new_file, result_file)
}
\arguments{
\item{old_file}{An excel file with old data.}

\item{new_file}{An excel file with new data.}

\item{result_file}{Name for output file.}
}
\value{
Modified excel file.
}
\description{
Function compares numererical data in worksheets and highlight cells that were changed. Function is limited to
specific regions in worksheets.
}
