% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_sector_denominator.R
\name{independent_sector_denominator}
\alias{independent_sector_denominator}
\title{Inpatient modified bed-days from Independent Sector Providers}
\usage{
independent_sector_denominator(data, bd_start_dt, bd_end_dt, admitted_dt,
  discharge_dt)
}
\arguments{
\item{bd_start_dt}{A date vector. The first day of the period of interest}

\item{admitted_dt}{A date vector. Date of patient admission}

\item{end_start_dt}{A date vector. The last day of the period of interest}

\item{end_start_dt}{A date vector. Date of patient discharge}
}
\value{
A dataframe containing the modified bed-days, number of days admitted and discharge for the period of interest
}
\description{
This function calculates inpatient modified bed-days as used in Public Health England's Independent Sector reports
}
\examples{

library(lubridate)
library(dplyr)

data <- data.frame(bd_start_dt = dmy("01/04/2017"),
                    bd_end_dt = dmy("31/03/2018"),
                    admitted_dt = dmy("17/03/2017"),
                    discharge_dt = dmy("01/04/2017"))

cbind(data, independent_sector_denominator(data, bd_start_dt, bd_end_dt, admitted_dt, discharge_dt))

data <- data.frame(
 admission_dt = c("17/03/2017","01/04/2017","01/04/2017","17/03/2017","01/04/2017","31/03/2018","23/04/2018","01/03/2018"),
 discharge_dt = c("02/04/2017","01/04/2017","03/04/2017","01/04/2018","23/04/2018","23/04/2018","23/04/2018","19/06/2018")
)

data$admission_dt <- dmy(data$admission_dt)
data$discharge_dt <- dmy(data$discharge_dt)
data$start_date <- dmy("01/04/2017")
data$end_date <- dmy("31/03/2018")

modified_bds <- independent_sector_denominator(data, bd_start_dt = start_date,
                                          bd_end_dt = end_date,
                                          admitted_dt = admission_dt,
                                          discharge_dt = discharge_dt)
cbind(data, modified_bds)

}
