% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpatient.R
\name{inpatient}
\alias{inpatient}
\title{inpatient}
\usage{
inpatient(patient_category, patient_location)
}
\arguments{
\item{patient_category}{character giving the admission category}

\item{patient_locationcharacter}{giving the patient location at time of specimen}
}
\value{
Numeric variable 1 for inpatients, 0 for other
}
\description{
Determines whether a record meets the criteria for an inpatient.
Will return 1 if the patient is an inpatient, day patient or emergency
assessment and is in an NHS acute trust at time of specimen.
}
\examples{
data(line_listing)

line_listing$inpatient <- inpatient(
    as.character(line_listing$patient_category),
    as.character(line_listing$patient_location))
}
