% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sch.R
\name{is_sch}
\alias{is_sch}
\title{Identifies specialist commissioning hubs}
\usage{
is_sch(x)
}
\arguments{
\item{x}{A string giving the CCG name}
}
\value{
A numeric vector giving 1 for a commissioning hub and zero for anything else.
}
\description{
Function to identify specialist commissioning hubs.
}
\examples{
x <- "YORKSHIRE AND HUMBER COMMISSIONING HUB"
is_sch(x)
x <- "NHS NEWCASTLE GATESHEAD CCG"
is_sch(x)
}
