% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_third_party.R
\name{is_third_party}
\alias{is_third_party}
\title{Check whether assignment is third party}
\usage{
is_third_party(x)
}
\arguments{
\item{x}{A numeric value from assignment method}
}
\value{
TRUE or FALSE
}
\description{
Check whether assignment is third party
}
\examples{
\dontrun{
is_third_party(9)
is_third_party(13)
}
}

