% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kh03_boilerplate.R
\name{kh03_boilerplate}
\alias{kh03_boilerplate}
\title{Adds KH03 boilerplate text columns to a data frame in preparation of kh03 data for upload to DCS}
\usage{
kh03_boilerplate(x)
}
\arguments{
\item{x}{A data frame to which columns will be added}
}
\value{
A data frame with new columns added
}
\description{
DCS denominator upload requires files in a certain format.
This is a function to add common variables to a data frame
}
\examples{
dat <- data.frame(trustcode = rep("AAA", 2), period = c(201415, 201516), stringsAsFactors = FALSE)
dat <- kh03_boilerplate(dat)
}
