% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kh03_year.R
\name{kh03_year}
\alias{kh03_year}
\title{Convert KH03-formatted dates to Mandatory data-formatted dates.}
\usage{
kh03_year(year, quarter, format)
}
\arguments{
\item{year}{The year column from KH03 data. String of 7 characters in format "yyyy-yy"}

\item{quarter}{A string giving the final month of the quarter, one of March, June, September, December}

\item{format}{A string giving the format desired for the output, one of fyear or cyear}
}
\value{
A numeric variable giving year and quarter
}
\description{
NHS England provide KH03 data periods as separate columns for financial year and quarter.
This converts to a format usable for the mandatory surveillance.
}
\examples{
kh03_year("2010-11", "March", "fyear")
kh03_year("2010-11", "March", "cyear")
kh03_year("2010-11", "June", "cyear")
kh03_year("2010-11", "June", "fyear")
kh03_year("2000-01", "March", "cyear")
}
