% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klebsiella_sp_group.R
\name{klebsiella_sp_group}
\alias{klebsiella_sp_group}
\title{Create group for Klebsiella species}
\usage{
klebsiella_sp_group(x)
}
\arguments{
\item{x}{A string giving the species name}
}
\value{
A string giving the reporting group
}
\description{
For reporting purposes, Klebsiella species are regrouped.
This function performs the regrouping.
Unrecognised species entries are given the value "This is not a known value".
Blank values are reclassified to 'Not speciated'
}
\examples{
test <- ""
klebsiella_sp_group(test)
}
