% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{line_listing}
\alias{line_listing}
\title{Data simulating output from the HCAI DCS line listing report}
\format{
A dataframe with 500 rows of 10 variables
\describe{
  \item{id}{Integer to simulate the record id}
  \item{collection}{A string giving the name of the organism collection}
  \item{age}{Patient age in years}
  \item{sex}{String giving patient sex}
  \item{specimen_date}{Date giving the date the specimen was taken}
  \item{onset_status}{String, whether hospital-onset or community-onset}
  \item{reporting_organisation_code}{A three-letter code for the reporting organisation}
  \item{ccg_code}{A three-letter code giving the attributed CCG}
  \item{pir_status}{A string giving the PIR status (only relevant to MRSA)}
  \item{prior_hc}{A string giving the prior healthcare group of the case (CDI only)}
  \item{patient_location}{A string giving location of patient at time of specimen}
  \item{patient_category}{A string giving admission status of patient at time of specimen}
}
}
\usage{
line_listing
}
\description{
The line-listing report of the HCAI DCS provides data on cases entered onto
the system. There is one row per case of infection. This dataset replicates
the data structure of the line listing report for a selection of variables
for use in testing and in examples.
}
\details{
All the data in the data set are simulated and do not represent actual
patient data.
}
\keyword{datasets}
