% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandatory_age.R
\name{mandatory_age}
\alias{mandatory_age}
\title{Calculate patient age in manner that conforms to mandatory surveillance requirements}
\usage{
mandatory_age(date_of_birth, specimen_date)
}
\arguments{
\item{date_of_birth}{Patient's date of birth in date format}

\item{specimen_date}{Date of specimen in date format}
}
\value{
Vector of integer ages
}
\description{
Calculates age as integer based on number of years lived.
Takes into account leap years.
}
\examples{
library(lubridate)
dob <- dmy("31/01/2000")
spec <- dmy("21/01/2001")
mandatory_age(dob, spec)
# Counter example :
round(as.numeric(spec-dob)/365.25, 1)
}

