% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandatory_age_group_bacteraemia.R
\name{mandatory_age_group_bacteraemia}
\alias{mandatory_age_group_bacteraemia}
\title{mandatory_age_group_bacteraemia(0)}
\usage{
mandatory_age_group_bacteraemia(x)
}
\description{
mandatory_age_group_bacteraemia(0)
}
