% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandatory_age_group_cdi.R
\name{mandatory_age_group_cdi}
\alias{mandatory_age_group_cdi}
\title{Mandatory age group for CDI cases
Calculates the mandatory age grouping based on numeric age}
\usage{
mandatory_age_group_cdi(x)
}
\arguments{
\item{x}{Age in years, calculated by \code{\link{mandatory_age}}
mandatory_age_group_cdi(0)
mandatory_age_group_cdi(2)}
}
\description{
Mandatory age group for CDI cases
Calculates the mandatory age grouping based on numeric age
}
