% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_everything.R
\name{merge_all}
\alias{merge_all}
\title{Essential function for merge_everything.}
\usage{
merge_all(merge_file, file_type, trust_code, ccg_code)
}
\arguments{
\item{merge_file}{Workbook to be merged.}

\item{file_type}{Non-MRSA (1) or MRSA (2) file type.}

\item{trust_code}{Codes for Trusts.}

\item{ccg_code}{Codes for CCG.}
}
\value{
Returns input for merge_everything function (workbook).
}
\description{
Operates on files, returns file.
}
