% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_everything.R
\name{merge_everything}
\alias{merge_everything}
\title{Merge annual tables after organisation mergers have occurred}
\usage{
merge_everything(merge_file, file_type, trust_list = NULL,
  ccg_list = NULL, result_file = "Result.xlsx")
}
\arguments{
\item{merge_file}{A path to an existing .xlsx file.}

\item{file_type}{Character string ("other", "mrsa") or corresponding numerical value (1, 2)}

\item{trust_list}{A list of Trust codes to be merged.}

\item{ccg_list}{A list of CCG codes to be merged.}
}
\value{
"Result.xlsx" file with applied merges.
}
\description{
Create an Excel file with applied mergers.
}
\details{
Function takes as the inputs: a path to the file, list(s) with Trust/CCG codes
and indication if this is MRSA/non-MRSA file. It utilizes these information to create new .xlsx file
that have applied mergers according to the codes provided. During this process all formatting
used in original file is lost. To use this function the right file must be used. Using wrong file type
could results in wrong calculations/errors. Structure of input code(s) is very important.
}
\note{
Some warnings are hidden with SupressWarnings() function.
}
\examples{
Excel file with the right structure must be used. Function works differently for MRSA and non-MRSA files:
MRSA_file <- "C:/My Documents/MRSA.xlsx"

Trust codes RCF, RFF and RAE will be merged into RCF and RF4, R1H into R1H.
MRSA_trust <- list(list(c("RCF", "RFF", "RAE"), "RCF"), list(c("RF4", "R1H"), "R1H"))

CCG codes 07L, 07M and 07N will be merged into 07L.
MRSA_ccg <- list(list(c("07L", "07M", "07N"), "07L"))

Execute function by, remember to choose the right file type:
merge_everything(MRSA_file, 2, MRSA_trust, MRSA_ccg)
}
