% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_everything.R
\name{merge_one}
\alias{merge_one}
\title{Essential function for merge_everything.}
\usage{
merge_one(worksheet_tbm, codes, is_SSRSS = FALSE)
}
\arguments{
\item{worksheet_tbm}{Worksheet to be merged.}

\item{codes}{Codes to be used.}

\item{is_SSRSS}{Indicator for MRSA special worksheets. Default is false.}
}
\value{
Returns input for merge_all function (worksheet).
}
\description{
Operates on worksheets, returns worksheet.
}
