% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_create_text_values.R
\name{mf_create_text_values}
\alias{mf_create_text_values}
\title{function to create text objects for monthly factsheets}
\usage{
mf_create_text_values(data = trend, collection, output_reqd)
}
\arguments{
\item{data}{A dataframe, in the context of the monthly factsheet, this will almost certainly be 'trend'}

\item{collection}{A character string, one of "mrsa", "mssa", "ecoli", "kleb", "paer", "cdi"}

\item{output_reqd}{A character string, one of "12_month", "3_month", "sum_12"}
}
\value{
A numeric value giving the percentage difference to different comparisons.
}
\description{
The factsheet creates object which contain percentage change in the count of cases for use within the bullet points of the text.
These bullet points compare the count of cases with the counts of cases at various different time points.
}
\examples{

data("mf_trend_data")
mf_trend_data <- mf_lag_trend(mf_trend_data)

cdi_12_month <- mf_create_text_values(data = mf_trend_data, collection = "cdi",
  output_reqd = "12_month")
cdi_12_month
}
