% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_fig1_fun.R
\name{mf_fig1_fun}
\alias{mf_fig1_fun}
\title{Monthly factsheet figure 1 function}
\usage{
mf_fig1_fun(data, collection)
}
\arguments{
\item{data}{A data frame giving the counts of cases for a month per row, with a date colummn which will be used as x variable and count of cases for a collection to be used as the y variable.}

\item{collection}{A string giving the name of the collection, which is the same as the name of the column in the data frame}
}
\value{
A ggplot showing the trend in counts of cases of the given collection
}
\description{
Function to produce figure 1 of the monthly factsheet.
}
\examples{

data(mf_trend_data)
head(mf_trend_data)

my_plot <- mf_fig1_fun(data = mf_trend_data, collection = "mrsa")
my_plot
my_plot <- mf_fig1_fun(data = mf_trend_data, collection = "mssa")
my_plot
my_plot <- mf_fig1_fun(data = mf_trend_data, collection = "ecoli")
my_plot
names(mf_trend_data)[5] <- "kleb"
my_plot <- mf_fig1_fun(data = mf_trend_data, collection = "kleb")
my_plot

}
