% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_lag_trend.R
\name{mf_lag_trend}
\alias{mf_lag_trend}
\title{lag mf trend data}
\usage{
mf_lag_trend(x = trend)
}
\arguments{
\item{x}{A data frame with columns for "cdi", "ecoli", "mrsa", "mssa",
"kleb", "paer" and "ecoli_ta".}
}
\value{
The same dataframe, but with new lagged variables added
}
\description{
Creates twelve month lag, three month sum, lagged by twelve months, and
twelve month sum, lagged by 12 months.
}
\examples{
data(mf_trend_data)
mf_trend_data <- mf_trend_data \%>\% mf_lag_trend()
head(mf_trend_data)
}
