% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_ccg_table.R
\name{mo_ccg_table}
\alias{mo_ccg_table}
\title{Monthly CCG table}
\usage{
mo_ccg_table(collection)
}
\arguments{
\item{collection}{A text string giving the data collection}
}
\value{
A wide data.frame with counts of cases per month in columns with CCGs as rows
}
\description{
Produce the monthly wide-format table for CCGs.
Takes long format data for CCGs and turns it wide with 13 months as variables.
Hard coded to operate on dataframe 'ccg_dat_all', it filters the data to a given collection, e.g. "MRSA" or "E. coli" as exported in the HCAI DCS line lists.
}
\examples{
data(monthly_ccg_data_raw)
ccg_dat_all <- monthly_ccg_data_raw
out_dat <- mo_ccg_table(collection = "MRSA")
head(out_dat)
}
