% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_curr_months.R
\name{mo_curr_months}
\alias{mo_curr_months}
\title{Create months for monthly outputs}
\usage{
mo_curr_months(x)
}
\arguments{
\item{x}{A date}
}
\value{
A vector containing 13 months
}
\description{
Use this to create a vector containing the 13 months covered by the mandatory surveillance monthly outputs.
}
\examples{
mo_curr_months(lubridate::today())
}
