% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_gnabsi_table.R
\name{mo_gnabsi_table}
\alias{mo_gnabsi_table}
\title{Monthly table for Gram-negative blood stream infection data}
\usage{
mo_gnabsi_table(data_fm, collection)
}
\arguments{
\item{data_fm}{A data frame to work on, usually either trust_dat_all or ccg_dat_all}

\item{collection}{A text string giving the data collection}
}
\value{
A wide data.frame with counts of cases by org and month.
}
\description{
Produce the monthly wide-format table for GNBSI .
Takes long format data for either trusts or CCGs and turns it wide with 13 months as variables.
The function will filter the data to a given collection, i.e. "K. pneumoniae", "E. coli" or "Pseudomonas" as exported in the HCAI DCS line lists.
}
\examples{
library(dplyr)
data(monthly_ccg_data_raw)
head(monthly_ccg_data_raw)
ccg_dat_all <- monthly_ccg_data_raw \%>\%
    rename(org_code = ccg_code, ho = apportioned) \%>\%
    mutate(data_collection = "E. coli", co = total_cases - ho)

my_out_dat <- mo_gnabsi_table(data_fm = ccg_dat_all, collection = "E. coli")

head(my_out_dat)
}
