% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_next_pub_date.R
\name{mo_next_pub_date}
\alias{mo_next_pub_date}
\title{Returns the next publication date of the monthly outputs
Used in construction of file paths to get to raw data and also for printing in text of monthly factsheet.}
\usage{
mo_next_pub_date(x)
}
\arguments{
\item{x}{A date, usually generated by lubridate::today()}
}
\value{
A data frame giving publication date, date for relevant data and month as a number for data
}
\description{
Returns the next publication date of the monthly outputs
Used in construction of file paths to get to raw data and also for printing in text of monthly factsheet.
}
\examples{
\dontrun{
next_pub <- mo_next_pub_date(lubridate::today())
}
}
