% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_prior_hc_table.R
\name{mo_prior_hc_table}
\alias{mo_prior_hc_table}
\title{mo_prior_hc_table}
\usage{
mo_prior_hc_table(dat, collection)
}
\arguments{
\item{dat}{a data frame containing at least following columns:
data_collection, org_code, year_no, month_no, total_cases, hoha, coha, coia,
coca, month_string}

\item{collection}{A string giving the data collection}
}
\value{
A dataframe with the table in long format
}
\description{
Create monthly table for prior healthcare exposures
}
\examples{
library(dplyr)
library(tidyr)
data(cdi_prior_hc_data)
mo_prior_hc_table(cdi_prior_hc_data)
}
