% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_pub_date.R
\name{mo_pub_date}
\alias{mo_pub_date}
\title{Nicely formats date of publication for monthly data}
\usage{
mo_pub_date(x = lubridate::today())
}
\arguments{
\item{x}{A date object}
}
\value{
A string object in format "Day, Month %d, YYYY"
}
\description{
The monthly tables have a string formatted date of publication in the header of each table.
The publication date is always the first Wednesday of the month in which publication occurs.
Data for month *m* are extracted and crunched in month *m* + 1, then published in month *m* + 2.
So, this function assumes that the user is working in *m& + 1 and will publish in the subsequent month.
This assumption can be over-ridden by supplying an actual date as an argument.
}
\examples{
mo_pub_date()
x <- lubridate::dmy("01/02/2015")
mo_pub_date(x)
}

