% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_tab_long.R
\name{mo_tab_long}
\alias{mo_tab_long}
\title{mo_tab_long}
\usage{
mo_tab_long(dat, collection_var, metric_var, org_var, year_var, month_var)
}
\arguments{
\item{dat}{A dataframe containing line-listing data}

\item{collection_var}{A string giving the data collection}

\item{metric_var}{Column in dat giving column for the metric}

\item{org_var}{The variable giving the organisation code}

\item{year_var}{The variable giving the calendar year}

\item{month_var}{The variable giving the calendar month}
}
\description{
Takes line list data and aggregates to long data for saving as long-format monthly tables.
Outputs are also sorted by organisation, time and metric.
}
