% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_ccg_data_raw}
\alias{monthly_ccg_data_raw}
\title{Data for monthly CCG table for MRSA}
\format{A data frame with 714 rows and 12 variables:
\describe{
    \item{data_collection}{String giving the data collection}
    \item{ccg_code}{String giving the ccg code}
    \item{fyear6}{String giving 6 character financial year}
    \item{year_no}{Numeric giving calendar year}
    \item{month_no}{Numeric giving month number}
    \item{total_cases}{Count of all cases for CCG and month}
    \item{apportioned}{Count giving apportioned cases for CCG and month}
    \item{nhs_trust_pir}{Count of trust-assigned PIR cases}
    \item{ccg_pir}{Count of CCG-assigned PIR cases}
    \item{third_party_pir}{Count of third party-assigned PIR cases}
    \item{fmonth}{Numeric giving financial year month, i.e. April = 1}
    \item{month_string}{Character string giving name of month}
}}
\usage{
monthly_ccg_data_raw
}
\description{
A sample data set for use with the mo_ccg_table function.
Is output in intermediate step in production of monthly tables and has counts of cases, by CCG and month in long format.
}
\keyword{datasets}
