% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsa_source_bacteraemia.R
\name{mrsa_source_bacteraemia}
\alias{mrsa_source_bacteraemia}
\title{Classify the source of bacteramia for MRSA}
\usage{
mrsa_source_bacteraemia(x)
}
\arguments{
\item{x}{Character variable giving the source of bacteraemia}
}
\value{
An ordered factor giving the grouped source
}
\description{
Groups the source of bacteraemia according to AEC table.
}
\examples{
mrsa_source_bacteraemia("CVC associated")
mrsa_source_bacteraemia("Endocarditis")
mrsa_source_bacteraemia("")
}
