% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_estimate.R
\name{nice_estimate}
\alias{nice_estimate}
\title{Formats estimates and 95\% confidence intervals for nice printing.}
\usage{
nice_estimate(estimate, lci, uci)
}
\arguments{
\item{estimate}{An estimate such as a rate ratio}

\item{lci}{The lower confidence interval}

\item{uci}{The upper confidence interval}
}
\value{
A string in format d.d (95\% CI: d.d-d.d)
}
\description{
Rounds estimates to 1 decimal place and copies similarly formatted confidence intervals inside brackets.
}
\examples{
nice_estimate(100.111, 90.0, 110.000002)
nice_estimate(0.9, 0.8001, 0.95)
}
