% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_names.R
\name{nice_names}
\alias{nice_names}
\title{Format variable names into standard format.}
\usage{
nice_names(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
Turns column names of data frames into standardised format.
Dots are replaced with underscores, trailing underscores are removed and everything is made lower case.
}
\examples{
data(mtcars)
names(mtcars) <- toupper(names(mtcars))
names(mtcars) <- nice_names(mtcars)
}
