% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_fin_qtr.R
\name{ordered_fin_qtr}
\alias{ordered_fin_qtr}
\title{Create factor variable from financial quarter}
\usage{
ordered_fin_qtr(data, quarter_in_long_format)
}
\arguments{
\item{data}{A dataframe}

\item{quarter_in_long_format}{A string variable giving long financial year
created using \code{\link{fy_long}}. Must be quoted.}
}
\value{
A dataframe with a new variable 'fin_qtr_ftr'
}
\description{
Creates factor variable from financial quarter to preserve correct ordering
in wide reshape of data.
}
\examples{
dat <- data.frame(sample_date = lubridate::dmy("01/01/2001"))
dat$sample_fy <- fy_long(dat$sample_date)
dat <- ordered_fin_qtr(dat, "sample_fy")
class(dat$fin_qtr_ftr)
levels(dat$fin_qtr_ftr)

dat <- data.frame(sample_date = lubridate::dmy("01/01/2001"))
dat$sample_fq <- fq_long(dat$sample_date)
dat <- ordered_fin_qtr(dat, "sample_fq")
class(dat$fin_qtr_ftr)
levels(dat$fin_qtr_ftr)

dat <- data.frame(sample_date = c(lubridate::dmy("01/01/2001"),
   lubridate::dmy("01/04/2001"),
   lubridate::dmy("01/07/2001"),
   lubridate::dmy("01/10/2001"),
   lubridate::dmy("01/01/2002")
   ))
dat$sample_fq <- fq_long(dat$sample_date)
dat <- ordered_fin_qtr(dat, "sample_fq")
class(dat$fin_qtr_ftr)
levels(dat$fin_qtr_ftr)

}
\seealso{
\code{\link{fy_long}}
}

