% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_m_y.R
\name{ordered_m_y}
\alias{ordered_m_y}
\title{Add ordered month and calendar year to a data frame}
\usage{
ordered_m_y(data, specimen_date)
}
\arguments{
\item{data}{A data frame}

\item{specimen_date}{A specimen date, in date format, as part of the data frame}
}
\value{
A dataframe with an additional column giving and ordered month in format "January 2015", ...
}
\description{
Function to add vector of ordered month-year factor to a dataframe.
To be used for preparation of monthly outputs.
}
\examples{
dat <- data.frame(spec_date = c(dmy("01/05/2015", "01/06/2015", "01/05/2015",
                                    "01/07/2015", "01/08/2015", "01/09/2015",
                                    "01/10/2015", "01/11/2015", "01/12/2015",
                                    "01/01/2016", "01/02/2016", "01/03/2016",
                                    "01/04/2016", "01/05/2016", "01/04/2015")),
                  stringsAsFactors = FALSE)
dat <- ordered_m_y(dat, "spec_date")
dat
head(dat$month_yr)
}
\seealso{
\code{\link{ordered_fin_qtr}}
}
