% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism_labeller.R
\name{organism_labeller}
\alias{organism_labeller}
\title{Create italicised organism names for plotting in facetted graphs}
\usage{
organism_labeller(labels)
}
\value{
A labeller object for use in a ggplot graph
}
\description{
Creating mixed normal font and italic font labels is tricky.
ggplot2 includes the \code{as_labeller} function and this function uses that to create labels for the organisms currently covered in the surveillance.
This function takes no arguments.
}
\examples{
  \dontrun{
  # These are only set as don't run so that the package builds more rapidly.
dat <- data.frame( variable = c("MRSA", "MSSA", "E. coli", "C. difficile",
    "Klebsiella spp", "Pseudomonas aeruginosa"),
    value = c(1, 2, 3, 4, 5, 6),
  stringsAsFactors = FALSE)
  ggplot2::ggplot(data = dat, ggplot2::aes(x = 1, y = value)) +
    ggplot2::facet_wrap( ~ variable, labeller = organism_labeller)
  }
}
