% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patloc_for_pir_assignment.R
\name{patloc_for_pir_assignment}
\alias{patloc_for_pir_assignment}
\title{Check whether a patient location is an NHS Trust or not}
\usage{
patloc_for_pir_assignment(x)
}
\arguments{
\item{x}{A string value giving the patient location}
}
\value{
"NHS Trust" or "CCG"
}
\description{
Not exported, called from assignment function.
}
\examples{
patloc_for_pir_assignment("NHS Acute Trust")
patloc_for_pir_assignment("Non-acute NHS Provider")
}

