% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popn_boilerplate.R
\name{popn_boilerplate}
\alias{popn_boilerplate}
\title{Adds population boilerplate text columns to a data frame in preparation of population denominator data for upload to DCS}
\usage{
popn_boilerplate(x)
}
\arguments{
\item{x}{A data frame to which columns will be added}
}
\value{
A data frame with new columns added
}
\description{
DCS denominator upload requires files in a certain format.
This is a function to add common variables to a data frame
}
\examples{
dat <- data.frame(trustcode = rep("AAA", 2), period = c(201415, 201516), stringsAsFactors = FALSE)
dat <- popn_boilerplate(dat)
}

