% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provis_orgname.R
\name{provis_orgname}
\alias{provis_orgname}
\title{Check whether provisional organisation is Trust or CCG}
\usage{
provis_orgname(x)
}
\arguments{
\item{x}{A numeric value from assignment method}
}
\value{
"Trust" or "CCG"
}
\description{
Check whether provisional organisation is Trust or CCG
}
\examples{
\dontrun{
provis_orgname("Cambridge University Hospitals NHS Foundation Trust")
provis_orgname("NHS Central London CCG")
}
}
