% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qec_age_table.R
\name{qec_age_table}
\alias{qec_age_table}
\title{Create age group table for QEC}
\usage{
qec_age_table(df, age_var, denom)
}
\arguments{
\item{df}{A data frame}

\item{age_var}{A variable giving the age group for the record}

\item{denom}{A variable giving the denominator for the age **group**}
}
\value{
a tibble containing the aggregated data
}
\description{
Produces rates (per 100,000) by age group for the QEC.
It's **really** important that the value for the denominator is the same across all records for the same age group
}
\examples{
dat <- data.frame(age_group = c(1, 1, 1, 2, 2, 2, 2),
                  denominator = c(500000,500000,500000, 600000, 600000, 600000, 600000))
qec_age_table(df = dat, age_var = age_group, denom = denominator)
}
