% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dcs.R
\name{query_dcs}
\alias{query_dcs}
\title{query dcs live}
\usage{
query_dcs(collection, cols = c(), date_from, date_to, reporting_org = "all",
  ccg = "all", test = FALSE)
}
\description{
query dcs live
}
\examples{

}
