% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_hcaidcs_package_manual.R
\name{show_hcaidcs_package_manual}
\alias{show_hcaidcs_package_manual}
\title{Display manual for this package in browser}
\usage{
show_hcaidcs_package_manual()
}
\description{
This function is Windows only.
A convenience function to create the pdf version of the help pages in the
**current working directory**.
Requires pdflatex to be installed on your machine.
The function takes no arguments and may run a little slowly.
}

