% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cap.R
\name{simple_cap}
\alias{simple_cap}
\title{Format text to mixed case capitalisation}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{A text string}
}
\description{
Formats text strings to mixed case text.
From https://stat.ethz.ch/R-manual/R-devel/library/base/html/chartr.html
}
\examples{
x <- "NHS ENGLAND NORTH"
simple_cap(x)
}

