% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subregion_test_data}
\alias{subregion_test_data}
\title{subregion_test_data}
\format{
A data frame with 14 observations on two columns
\describe{
\item{ODS_CD}{ODS code of subregion for joining}
\item{rand_val}{A random value for filling the areas}
}
}
\usage{
subregion_test_data
}
\description{
A data frame with random data for plotting on subregions
}
\keyword{datasets}
