% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subregions_sp_df}
\alias{subregions_sp_df}
\title{NHS subregion geographies for England}
\format{A data frame with ~71.5k observations and 16 columns:
\describe{
 \item{long}{Longitude}
 \item{lat}{Lattitude}
 \item{group}{Group id for polygons}
 \item{id}{Numeric id for subregion, 0-9}
 \item{GSS_CD}{ONS code for subregion}
 \item{GSS_CD}{ONS name for subregion}
 \item{ODS_CD}{HSCIC ODS code for subregion}
 \item{ODS_P_CD}{HSCIC ODS code for parent region}
 \item{ODS_P_CD}{ONS code for parent region}
}}
\source{
Office for National Statistics
}
\usage{
subregions_sp_df
}
\description{
A dataframe containing the geographies for NHS subregions.
}
\keyword{datasets}

