% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_period2fyear.R
\name{time_period2fyear}
\alias{time_period2fyear}
\title{Convert time period variable to fyear for merging with KH03 data}
\usage{
time_period2fyear(x)
}
\arguments{
\item{x}{A text string, in format "qqq-qqq yyyy"}
}
\value{
A text string giving the quarter in fyear format.
}
\description{
HCAI DCS QMLR data sets include a time period variable in the format "qqq-qqq yyyy".
To merge with KH03 data, this needs to be converted to "yyyyq".
For fyear, q1 is April-June and q4 is Jan-Feb.
}
\examples{
x <- "Oct-Dec 2015"
time_period2fyear(x)
}

