% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_onset.R
\name{time_to_onset}
\alias{time_to_onset}
\title{Calculates time to onset for inpatients}
\usage{
time_to_onset(date_admitted, date_specimen, patient_location)
}
\arguments{
\item{date_admitted}{Admission date in date format}

\item{date_specimen}{Date specimen taken in date format}

\item{patient_location}{Location of patient at time of specimen. One of:
"In-patient", "Day patient" or "Emergency Assessment"}
}
\value{
A numeric value
}
\description{
Only applicable for inpatients. Where date of specimen is before the date of
admission, \code{NA} will be returned.
}
\examples{
onset <- dmy("05-01-2011")
admitted <- dmy("01-01-2011")
patient_loc <- "In-patient"
time_to_onset(admitted, onset, patient_loc)
\dontrun{
## Should return NA
onset <- NA
time_to_onset(admitted, onset, patient_loc)
patient_loc <- ""
onset <- dmy("05-01-2011")
time_to_onset(admitted, onset, patient_loc)
}
}

