% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_orgs_dcs.R
\name{update_independent}
\alias{update_independent}
\title{Update independent sector providers and independent sector sites.}
\usage{
update_independent(organisation,
  input_file_path = "K:/Mandatory databases/maintaining_new_org_structure_new/ods_files/",
  output_file_path = "K:/Mandatory databases/maintaining_new_org_structure_new/ods_files/updates/")
}
\arguments{
\item{organisation}{A character giving the organisation to be updated.}

\item{input_file_path}{A character string giving the path to the ods csv files, but without specifying the csv file itself.}

\item{output_file_path}{A character string giving the path where outputs will be saved.}
}
\value{
Four files, two csv and two zip for each ephp and ephpsite files.
}
\description{
This function takes two csv files: ephp.csv and ephpsite.csv limits data to the specified organisation(s) then exports to zip files.
The returned zip files can be used to upload to the HCAI DCS to update organisation mappings.
}
\examples{
\dontrun{
 update_independent("AXG")
}
}
