% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann_tab_long.R
\name{ann_tab_long}
\alias{ann_tab_long}
\title{Create long annual tables from line listing data}
\usage{
ann_tab_long(dat, collection_var, collection_string, org_code, metric_type,
  metric_var, output, spec_date)
}
\arguments{
\item{dat}{A dataframe containing line-listing data}

\item{collection_var}{A string giving the data collection}

\item{collection_string}{A string giving the data collection}

\item{org_code}{The variable giving the organisation code}

\item{metric_type}{A string giving the metric to be tabulated, one of "onset", "pir" or "prior_hc"}

\item{metric_var}{Column in dat giving column for the metric}

\item{output}{A string giving the output required. Either an annual table or quarterly}

\item{spec_date}{Column in dat giving specimen date}
}
\value{
A dataframe with counts of cases in long format, by organisation and time period
}
\description{
From FY 2018/19, annual tables will be moving to a long format rather than a
wide format. This function takes line-listing data and aggregates by
organisation and time period, producing counts of cases by a given metric
(apportioning or prior healthcare exposure).
}
\details{
The data are expanded to cover time periods in the data, but there may well
be organisations that are not in the line-listing data that will be merged in
later to make a complete table.

In addition, PHEC and organisation names will need to be merged in, as will
denominator data, and rates will need to be calculated.
}
\examples{
#library(dplyr)
data(line_listing)
line_listing$fy_long <- fy_long(line_listing$specimen_date)

ann_tab_long(dat = line_listing, collection_var = collection,
  collection_string = "C. difficile", org_code = reporting_organisation_code,
  metric_type = "prior_hc", metric_var = prior_hc, output = "annual",
  spec_date = specimen_date)
}
