% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_algorithm.R
\name{assignment_algorithm}
\alias{assignment_algorithm}
\title{Replicate the PIR assignment algorithm}
\usage{
assignment_algorithm(pircasestatus, assignmentmethodcode, patientlocation,
  provisionalorganisationname, finalpirassignedorganisation)
}
\arguments{
\item{pircasestatus}{Text string of either 'Final assignment' or 'Provisional assignment'}

\item{assignmentmethodcode}{Numeric variable giving assignment method}

\item{patientlocation}{Text string giving patient location}

\item{provisionalorganisationname}{Text string giving the name of the provisionally assigned organisation}

\item{finalpirassignedorganisation}{Text string giving name of finally assigned pir organisation}
}
\value{
string giving the name of the finally assigned organisation or one of "NHS Trust" or "CCG"
}
\description{
Replicate the PIR assignment algorithm
}
\examples{
dat$new <- assignment_algorithm(dat$pircasestatus, dat$assignmentmethodcode,
           dat$patientlocation, dat$provisionalorganisationname,
           dat$finalpirassignedorganisation)
dat
}

