% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_algorithm.R
\name{assignment_algorithm}
\alias{assignment_algorithm}
\title{Replicate the PIR assignment algorithm}
\usage{
assignment_algorithm(pircasestatus, assignmentmethodcode, patientlocation,
  provisionalorganisationname, finalpirassignedorganisationtype)
}
\arguments{
\item{pircasestatus}{Text string of either 'Final assignment' or 'Provisional assignment'}

\item{assignmentmethodcode}{Numeric variable giving assignment method}

\item{patientlocation}{Text string giving patient location}

\item{provisionalorganisationname}{Text string giving the name of the provisionally assigned organisation}

\item{finalpirassignedorganisationtype}{Text string giving organisation type of finally assigned pir organisation. One of either "NHS Trust" or "Clinical Commissioning Group"}
}
\value{
string giving the name of the finally assigned organisation or one of "NHS Trust" or "CCG"
}
\description{
Provisionally assigned cases are counted as if final assignment status, i.e.
a case provisionally assigned to a Trust will be reported as Trust-assigned
in any report.
}
\examples{
dat <- structure(list(
  pircasestatus = c("Final assignment", "Final assignment","Final assignment",
                    "Final assignment", "Provisional assignment",
                    "Provisional assignment", "Provisional assignment",
                    "Provisional assignment", "Provisional assignment",
                    "Provisional assignment", "Provisional assignment"),
  assignmentmethodcode = c(9L, 10L, 10L, 8L, 10L, 10L, 13L, 15L, 15L, 11L, 11L),
  patientlocation = c("NHS Acute Trust", "NHS Acute Trust",
                      "Independent Sector Provider",
                      "Independent Sector Provider",
                                          "NHS Acute Trust",
                      "Independent Sector Provider", "NHS Acute Trust",
                                          "NHS Acute Trust",
                      "NHS Acute Trust", "NHS Acute Trust", "NHS Acute Trust"),
  provisionalorganisationname = c(
    "Cambridge University Hospitals NHS Foundation Trust",
    "Cambridge University Hospitals NHS Foundation Trust",
    "Cambridge University Hospitals NHS Foundation Trust",
    "Cambridge University Hospitals NHS Foundation Trust",
    "Cambridge University Hospitals NHS Foundation Trust",
    "NHS Central London CCG",
    "Cambridge University Hospitals NHS Foundation Trust",
    "Cambridge University Hospitals NHS Foundation Trust",
    "NHS Central London CCG",
    "Cambridge University Hospitals NHS Foundation Trust",
    "NHS Central London CCG"),
  finalpirassignedorganisationtype = c(
    "NHS Trust",
    "NHS Trust",
    "Clinical Commissioning Group",
    "Clinical Commissioning Group",
    "NHS Trust",
    "NHS Trust",
    "NHS Trust",
    "NHS Trust",
    "NHS Trust",
    "NHS Trust",
    "NHS Trust")),
  .Names = c("pircasestatus", "assignmentmethodcode", "patientlocation",
             "provisionalorganisationname", "finalpirassignedorganisation"),
                 class = "data.frame", row.names = c(NA, -11L))

dat$new <- assignment_algorithm(dat$pircasestatus,
         dat$assignmentmethodcode, dat$patientlocation, dat$provisionalorganisationname,
           dat$finalpirassignedorganisation)
dat
}
