% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_algorithm.R
\name{assignment_algorithm}
\alias{assignment_algorithm}
\title{Replicate the PIR assignment algorithm}
\usage{
assignment_algorithm(pircasestatus, assignmentmethodcode, patientlocation,
  patientcategory, provisionalorganisationname,
  finalpirassignedorganisationtype)
}
\arguments{
\item{pircasestatus}{Text string of either 'Final assignment' or 'Provisional assignment'}

\item{assignmentmethodcode}{Numeric variable giving assignment method}

\item{patientlocation}{Text string giving patient location}

\item{patientcategory}{Text string giving patient inpatient status (one of "", "A&E only", "Day patient", "Emergency Assessment" "In-patient", "Other", "Outpatient", "Regular Attender", "Unknown" )}

\item{provisionalorganisationname}{Text string giving the name of the provisionally assigned organisation}

\item{finalpirassignedorganisationtype}{Text string giving organisation type of finally assigned pir organisation. One of either "NHS Trust" or "Clinical Commissioning Group"}
}
\value{
string giving the name of the finally assigned organisation or one of "NHS Trust" or "CCG"
}
\description{
Provisionally assigned cases are counted as if final assignment status, i.e.
a case provisionally assigned to a Trust will be reported as Trust-assigned
in any report.
}
\examples{

data(assignment_data)
assignment_data$new <- assignment_algorithm(pircasestatus = assignment_data$pircasestatus,
        assignmentmethodcode = assignment_data$assignmentmethodcode,
        patientlocation = assignment_data$patientlocation,
        patientcategory = assignment_data$patient_category,
        provisionalorganisationname = assignment_data$provisionalorganisationname,
        finalpirassignedorganisationtype = assignment_data$finalpirassignedorganisation)
assignment_data

dat2 <- assignment_data[13, ]
dat2$new <- NULL
dat2$new <- assignment_algorithm(pircasestatus = dat2$pircasestatus,
        assignmentmethodcode = dat2$assignmentmethodcode,
        patientlocation = dat2$patientlocation,
        patientcategory = dat2$patient_category,
        provisionalorganisationname = dat2$provisionalorganisationname,
        finalpirassignedorganisationtype = dat2$finalpirassignedorganisation)
dat2

data(assignment_data2)
assignment_data2$new <- assignment_algorithm(pircasestatus = assignment_data2$pir_case_status,
        assignmentmethodcode = assignment_data2$assignment_method_code,
        patientlocation = assignment_data2$patient_location,
        patientcategory = assignment_data2$patient_category,
        provisionalorganisationname = assignment_data2$provisional_organisation_name,
        finalpirassignedorganisationtype = assignment_data2$final_pir_assigned_organisation_type)
assignment_data2
}
