% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assignment_data2}
\alias{assignment_data2}
\title{Data for testing assignement algorithm}
\format{
A data frame with 15 rows of 7 variables:
\describe{
 \item{pir_case_status}{string}
 \item{assignment_method_code}{Numeric giving assignment method}
 \item{patient_location}{string}
 \item{patient_category}{string}
 \item{provisional_organisation_name}{string}
 \item{final_pir_assigned_organisation_type}{string}
 \item{id}{A numeric variable giving an anonymous HCAIDCS case id}
}
}
\usage{
assignment_data2
}
\description{
A sample data set for testing the assignment_algorithm function.
}
\keyword{datasets}
