% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdi_prior_hc_data}
\alias{cdi_prior_hc_data}
\title{Data for testing assignement algorithm}
\format{A data frame with 14 rows of 6 variables:
\describe{
 \item{pircasestatus}{string}
 \item{assignmentmethodcode}{Numeric giving assignment method}
 \item{patientlocation}{string}
 \item{provisionalorganisationname}{string}
 \item{finalpirassignedorganisation}{string}
 \item{patient_category}{string}
}}
\usage{
cdi_prior_hc_data
}
\description{
A sample data set for testing the assignment_algorithm function.
}
\keyword{datasets}
