% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Gives the absolute value of the per cent or total change between two values}
\usage{
change(from, to, type)
}
\arguments{
\item{from}{The starting value}

\item{to}{The finishing value}

\item{type}{Either "percent" or "value"}
}
\description{
Gives the absolute value of the per cent or total change between two values
}
\examples{
change(200, 100, "percent")
change(200, 100, "value")
change(100, 200, "percent")
}
