% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_csv2r.R
\name{convert_csv2RCCG}
\alias{convert_csv2RCCG}
\title{Convert .csv file to global variable that can be used in merge_everything function.}
\usage{
convert_csv2RCCG(ccg_file)
}
\arguments{
\item{ccg_file}{A file with CCG codes.}
}
\value{
R object with the structure: \cr
(list(list(old codes 1), new code 1), list((old codes 2), new code 2), ...) \cr
}
\description{
Creates ccg_list object from ".csv" file. Csv files must contain only CCG codes to be merged.
This file needs to be formatted using following specification:
\itemize{
\item first row contains only column names, such as "old codes", "new code" etc.
\item "odd" columns (1, 3, 5...) contains old codes in a separate cells
\item "even" columns (2, 4, 6...) contains new code in 1 cell only
\item "odd" column 1 corresponds to "even" column 2, 3 to 4, 5 to 6 ...
\item there should be no empty spaces between the cells with data
\item there should be only data necessary for conversion, for example to convert 2 merge events with 3 and 4 codes
you will need to use csv with data in 4 columns, where first row is description only (no codes). First column
would have 4 rows (description + 3 codes), second column 2 rows (description + new code), third column 5 rows (description
+ 4 codes) and final fourth column 2 rows(description + new code)
}
For example how properly format file please see  "test_convertcsv2rCCG.csv" file in the main directory
of HCAIDCS package.
}
