% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_compare.R
\name{data_compare}
\alias{data_compare}
\title{Compare two data frames to identify cells which are different}
\usage{
data_compare(currentdf, lastdf)
}
\arguments{
\item{currentdf}{A data frame in wide format giving the values for the current publication.}

\item{lastdf}{A data frame in wide format giving the values for the most recent publication prior to the one being prepared.}
}
\value{
A matrix specifying cells that have changed.
}
\description{
Used to identify cells which have changed between publications for later highlighting in Excel export.
With thanks to Amy Mikhail for writing this function for me.
}
\examples{
last_pub <- data.frame(org = c("a", "b", "c"),
jan16 = sample(1:100, 3, replace=T),
feb16 = sample(1:100, 3, replace=T),
mar16 = sample(1:100, 3, replace=T),
jan17 = sample(1:100, 3, replace=T))

# copy the last pub, but drop the oldest month
this_pub <- last_pub[c(1,3:5)]
# add new month
this_pub$feb17 <- sample(1:100, 3, replace=T)
# change one value only
this_pub$mar16[2] <- this_pub$mar16[2] + 1
# Checki it out:
this_pub
highlight <- data_compare(this_pub, last_pub)
}
