% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcs_specialty_grouping.R
\name{dcs_speciality_grouping}
\alias{dcs_speciality_grouping}
\title{Group specialties from the HCAI DCS}
\usage{
dcs_speciality_grouping(spec_nm)
}
\arguments{
\item{spec_nm}{Character. Specialty name}
}
\description{
Group specialties from the HCAI DCS
}
\examples{

# The function categorises specialties into groups by string matching on `spec_nm`.
# Numbers will be compared as characters
# NAs of all types will evaluate to "not available"
# Any value supplied that does not meet any of the string matching conditions evaluate to "others"

dcs_speciality_grouping("Clinical oncology (radiotherapy)")

dcs_speciality_grouping("PIC Gastroenterology")

dcs_speciality_grouping("PIC Nephrology")

dcs_speciality_grouping("")

dcs_speciality_grouping("  ")

dcs_speciality_grouping(NA)

dcs_speciality_grouping("xds")

dcs_speciality_grouping(1)

dcs_speciality_grouping(2.1)

dcs_speciality_grouping(-2.1)


}
\author{
Olisaeloka Nsonwu
}
