% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_wednesday.R, R/mo_next_pub_date.R
\name{first_wednesday}
\alias{first_wednesday}
\alias{first_wednesday}
\title{First Wednesday}
\usage{
first_wednesday(x)

first_wednesday(x)
}
\arguments{
\item{x}{A date}

\item{x}{A date in Year, Month, Day format}
}
\value{
A date giving the first Wednesday of the month.

A date object
}
\description{
Returns the first Wednesday of a month as a date in YYYY-MM-DD.
Used in monthly outputs for determining publication date.
Relies on lubridate for date calculations.

Returns the first Wednesday of a month
Used to calculated the next publication date of monthly outputs.
}
\examples{
x <- lubridate::dmy("01/01/1970")
first_wednesday(x)
weekdays(first_wednesday(x))
first_wednesday(lubridate::floor_date(lubridate::today(), "month"))
\dontrun{
first_wednesday(lubridate::floor_date(today(), "month") \%m+\%
    lubridate::months(1))
    }
}
