% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_wednesday.R
\name{first_wednesday}
\alias{first_wednesday}
\title{First Wednesday}
\usage{
first_wednesday(x)
}
\arguments{
\item{x}{A date}
}
\value{
A date giving the first Wednesday of the month.
}
\description{
Returns the first Wednesday of a month as a date in YYYY-MM-DD.
Used in monthly outputs for determining publication date.
Relies on lubridate for date calculations.
}
\examples{
first_wednesday(lubridate::floor_date(lubridate::today(), "month"))
\dontrun{
first_wednesday(lubridate::floor_date(today(), "month") \%m+\%
    lubridate::months(1))
    }
x <- as.Date("01/01/2017", format = "\%d/\%m/\%Y")
weekdays(first_wednesday(x))
}
