% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandatory_age_group.R
\name{mandatory_age_group}
\alias{mandatory_age_group}
\title{Create age group from age}
\usage{
mandatory_age_group(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A factor vector with levels <1 1-14 15-44 45-64 65-74 75-84 ge85
}
\description{
Creates age group as ordered factor according to standard ages for HCAIDCS.
Values are rounded down using floor before grouping.
}
\examples{
mandatory_age_group(0)
mandatory_age_group(0.038)
mandatory_age_group(1)
mandatory_age_group(14)
mandatory_age_group(15)
mandatory_age_group(seq(1,100, 1))
}
\seealso{
\code{\link{mandatory_age}}
}

