% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandatory_age_group_bacteraemia.R
\name{mandatory_age_group_bacteraemia}
\alias{mandatory_age_group_bacteraemia}
\title{mandatory_age_group_bacteraemia(0)}
\usage{
mandatory_age_group_bacteraemia(x)
}
\arguments{
\item{x}{Age as a numeric value}
}
\value{
An ordered factor giving the age group
}
\description{
Sub-function for creating mandatory age groups for bacteraemia cases
}
