% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandatory_age_group_cdi.R
\name{mandatory_age_group_cdi}
\alias{mandatory_age_group_cdi}
\title{mandatory_age_group_cdi(0)
mandatory_age_group_cdi(2)}
\usage{
mandatory_age_group_cdi(x)
}
\description{
mandatory_age_group_cdi(0)
mandatory_age_group_cdi(2)
}
