% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_everything.R
\name{merge_everything}
\alias{merge_everything}
\title{Merge annual tables after organisation mergers have occurred}
\usage{
merge_everything(merge_file, file_type, trust_list = NULL,
  ccg_list = NULL, result_file = "Result.xlsx")
}
\arguments{
\item{merge_file}{A path to an existing .xlsx file.}

\item{file_type}{Character string ("other", "mrsa") or corresponding numerical value (1, 2)}

\item{trust_list}{A list of Trust codes to be merged.}

\item{ccg_list}{A list of CCG codes to be merged.}

\item{result_file}{Name for output file in "name.xlsx" format.}
}
\value{
"Result.xlsx" file with applied merges.
}
\description{
Create an Excel file with applied mergers.
}
\details{
Function takes as the inputs: a path to the file, list(s) with Trust/CCG codes
and indication if this is MRSA/non-MRSA file. It utilizes these information to create new .xlsx file
that have applied mergers according to the codes provided. During this process all formatting
used in original file is lost. To use this function the right file must be used. Using wrong file type
could results in wrong calculations/errors. Structure of input code(s) is very important.
}
\note{
Some warnings are hidden with SupressWarnings() function.
}
