% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_everything.R
\name{merge_one}
\alias{merge_one}
\title{Essential function for merge_everything.}
\usage{
merge_one(worksheet_tbm, codes, is_SSRSS = FALSE)
}
\description{
Operates on worksheets, returns worksheet.
}
