% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_m_y.R
\name{ordered_m_y}
\alias{ordered_m_y}
\title{Add ordered month and calendar year to a data frame}
\usage{
ordered_m_y(data, specimen_date)
}
\arguments{
\item{data}{A data frame}

\item{specimen_date}{A specimen date, in date format, as part of the data frame}
}
\value{
A dataframe with an additional column giving and ordered month in format "January 2015", ...
}
\description{
Function to add vector of ordered month-year factor to a dataframe.
To be used for preparation of monthly outputs.
}
\examples{
\dontrun{
# This is failing R cmd check because can't find dat and I don't understand why
# so excluding from examples for now. This should run fine if copied-and-pasted
ordered_m_y(data = dat, specimen_date = spec_date)
}
}
\seealso{
\code{\link{ordered_fin_qtr}}
}
