% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subregions_sp_df}
\alias{subregions_sp_df}
\title{NHS subregion geographies for England}
\format{A data frame with ~71.5k observations and 16 columns:
\describe{
 \item{long}{Longitude}
 \item{lat}{Lattitude}
 \item{OBJECTID}{Id for an object. Only used for plotting}
 \item{hole}{Logical value for whether geography is a hole. Only used for plotting}
 \item{piece}{Only used for plotting}
 \item{order}{Gives plotting order}
 \item{group}{Group id for polygons}
 \item{id}{Numeric id for subregion, 0-9}
 \item{GSS_CD}{ONS code for subregion}
 \item{GSS_NM}{ONS name for subregion}
 \item{Shape_STAr}{Only used for plotting}
 \item{Shape_STLe}{Only used for plotting}
 \item{centroid_lat}{Latitude values for subregion centroids for plotting labels}
 \item{centroid_long}{Longitude values for subregion centroids for plotting labels}
 \item{ONS_P_NM}{ONS Parent Name}
 \item{ODS_CD}{ODS ODS code for subregion}
 \item{ODS_P_CD}{ODS ODS code for parent region}
 \item{ONS_P_CD}{ONS Parent Code}
}}
\source{
Office for National Statistics
}
\usage{
subregions_sp_df
}
\description{
A dataframe containing the geographies for NHS subregions.
}
\keyword{datasets}
