% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_onset.R
\name{time_to_onset}
\alias{time_to_onset}
\title{Calculates time to onset for inpatients}
\usage{
time_to_onset(date_admitted, date_specimen, patient_category,
  patient_location)
}
\arguments{
\item{date_admitted}{Admission date in date format}

\item{date_specimen}{Date specimen taken in date format}

\item{patient_category}{Category of patient at time of specimen. One of:
"In-patient", "Day patient" or "Emergency Assessment"}

\item{patient_location}{Location of patient at time of specimen}
}
\value{
A numeric value
}
\description{
Only applicable for inpatients. Where date of specimen is before the date of
admission, \code{NA} will be returned.
}
\examples{
onset <- lubridate::dmy("05-01-2011")
admitted <- lubridate::dmy("01-01-2011")
patient_cat <- "In-patient"
patient_loc <- "NHS Acute Trust"
time_to_onset(admitted, onset, patient_cat, patient_loc)
dat <- data.frame(onset = onset, admitted = admitted, patient_cat = patient_cat,
patient_loc = patient_loc)
dat$tto <- time_to_onset(dat$admitted, dat$onset, dat$patient_cat, dat$patient_loc)
\dontrun{
## Should return NA
onset <- NA
time_to_onset(admitted, onset, patient_cat, patient_loc)
patient_cat <- ""
onset <- lubridate::dmy("05-01-2011")
time_to_onset(admitted, onset, patient_cat)
time_to_onset(date_admitted = lubridate::dmy("01-01-2011"),
 date_specimen = lubridate::dmy("05-01-2011"),
 patient_category = "Outpatient", patient_location = "NHS Acute Trust")
}
}
